add likesControllers directory with atomic controllers for async requests
