+++ src/controllers/usersControllers/user.controller.js
@@ -1,4 +1,4 @@
-import * as user from '../usecases/user.usecase.js'
+import * as user from '../../usecases/user.usecase.js'
 
 const getUsers = async (req, res) => {
   try {
